Split ATTRS into separate files; update tests
