+++ server/src/packages/packages.controller.ts
@@ -3,11 +3,12 @@ import { PackagesService } from './packages.service';
 import { JsonApiSerializeInterceptor } from '../json-api-serialize.interceptor';
 import { JsonApiDeserializePipe } from '../json-api-deserialize.pipe';
 import { AuthenticateGuard } from '../authenticate.guard';
-import { PAS_FORM_ATTRS } from './pas-form/pas-form.controller';
 import { APPLICANT_ATTRS } from './pas-form/applicants/applicants.controller';
 import { BBL_ATTRS } from './pas-form/bbls/bbls.controller';
 import { pick } from 'underscore';
 import { PROJECT_ATTRS } from '../projects/projects.controller';
+import { RWCDS_FORM_ATTRS } from './rwcds-form/rwcds-form.attrs';
+import { PAS_FORM_ATTRS } from './pas-form/pas-form.attrs';
 
 export const PACKAGE_ATTRS = [
   'statuscode',
@@ -56,47 +57,6 @@ export const PAS_FORM_PROJECTADDRESS_ATTRS = [
   'dcp_validatedaddressoverride',
   'dcp_addressvalidateddate',
   'statecode',
-]
-
-export const RWCDS_FORM_ATTRS = [
-  'dcp_describethewithactionscenario',
-  'dcp_isplannigondevelopingaffordablehousing',
-  'dcp_includezoningtextamendment',
-  'dcp_existingconditions',
-  'processid',
-  'statecode',
-  'importsequencenumber',
-  'versionnumber',
-  'dcp_rationalbehindthebuildyear',
-  'createdon',
-  'modifiedon',
-  'dcp_isapplicantseekingaction',
-  'dcp_whichactionsfromotheragenciesaresought',
-  'dcp_proposedprojectdevelopmentdescription',
-  'dcp_version',
-  'dcp_projectsitedescription',
-  'dcp_sitehistory',
-  'dcp_purposeandneedfortheproposedaction',
-  'dcp_describethenoactionscenario',
-  'dcp_applicant',
-  'dcp_hasprojectchangedsincesubmissionofthepas',
-  'traversedpath',
-  'statuscode',
-  'dcp_borough',
-  'dcp_projectname',
-  'dcp_rwcdsexplanation',
-  'dcp_communitydistrict',
-  'dcp_howdidyoudeterminethenoactionscenario',
-  'dcp_name',
-  'dcp_isrwcdsscenario',
-  'timezoneruleversionnumber',
-  'dcp_howdidyoudeterminethiswithactionscena',
-  'dcp_buildyear',
-  'dcp_developmentsiteassumptions',
-  'dcp_constructionphasing',
-  'dcp_date',
-  'overriddencreatedon',
-  'utcconversiontimezonecode',
 ]
 
 @UseInterceptors(new JsonApiSerializeInterceptor('packages', {
