+++ server/src/packages/rwcds-form/rwcds-form.attrs.ts
@@ -1,0 +1,40 @@
+export const RWCDS_FORM_ATTRS = [
+  'dcp_describethewithactionscenario',
+  'dcp_isplannigondevelopingaffordablehousing',
+  'dcp_includezoningtextamendment',
+  'dcp_existingconditions',
+  'processid',
+  'statecode',
+  'importsequencenumber',
+  'versionnumber',
+  'dcp_rationalbehindthebuildyear',
+  'createdon',
+  'modifiedon',
+  'dcp_isapplicantseekingaction',
+  'dcp_whichactionsfromotheragenciesaresought',
+  'dcp_proposedprojectdevelopmentdescription',
+  'dcp_version',
+  'dcp_projectsitedescription',
+  'dcp_sitehistory',
+  'dcp_purposeandneedfortheproposedaction',
+  'dcp_describethenoactionscenario',
+  'dcp_applicant',
+  'dcp_hasprojectchangedsincesubmissionofthepas',
+  'traversedpath',
+  'statuscode',
+  'dcp_borough',
+  'dcp_projectname',
+  'dcp_rwcdsexplanation',
+  'dcp_communitydistrict',
+  'dcp_howdidyoudeterminethenoactionscenario',
+  'dcp_name',
+  'dcp_isrwcdsscenario',
+  'timezoneruleversionnumber',
+  'dcp_howdidyoudeterminethiswithactionscena',
+  'dcp_buildyear',
+  'dcp_developmentsiteassumptions',
+  'dcp_constructionphasing',
+  'dcp_date',
+  'overriddencreatedon',
+  'utcconversiontimezonecode',
+];
