+++ server/src/packages/rwcds-form/rwcds-form.controller.spec.ts
@@ -2,6 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { RwcdsFormController } from './rwcds-form.controller';
 import { CrmModule } from '../../crm/crm.module';
 import { RwcdsFormService } from './rwcds-form.service';
+import { PasFormService } from '../pas-form/pas-form.service';
 
 describe('RwcdsForm Controller', () => {
   let controller: RwcdsFormController;
@@ -9,7 +10,7 @@ describe('RwcdsForm Controller', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [CrmModule],
-      providers: [RwcdsFormService],
+      providers: [RwcdsFormService, PasFormService],
       controllers: [RwcdsFormController],
     }).compile();
