+++ server/src/packages/rwcds-form/rwcds-form.controller.ts
@@ -4,47 +4,7 @@ import { JsonApiSerializeInterceptor } from '../../json-api-serialize.intercepto
 import { JsonApiDeserializePipe } from '../../json-api-deserialize.pipe';
 import { RwcdsFormService } from './rwcds-form.service';
 import { pick } from 'underscore';
-
-export const RWCDS_FORM_ATTRS = [
-  'dcp_describethewithactionscenario',
-  'dcp_isplannigondevelopingaffordablehousing',
-  'dcp_includezoningtextamendment',
-  'dcp_existingconditions',
-  'processid',
-  'statecode',
-  'importsequencenumber',
-  'versionnumber',
-  'dcp_rationalbehindthebuildyear',
-  'createdon',
-  'modifiedon',
-  'dcp_isapplicantseekingaction',
-  'dcp_whichactionsfromotheragenciesaresought',
-  'dcp_proposedprojectdevelopmentdescription',
-  'dcp_version',
-  'dcp_projectsitedescription',
-  'dcp_sitehistory',
-  'dcp_purposeandneedfortheproposedaction',
-  'dcp_describethenoactionscenario',
-  'dcp_applicant',
-  'dcp_hasprojectchangedsincesubmissionofthepas',
-  'traversedpath',
-  'statuscode',
-  'dcp_borough',
-  'dcp_projectname',
-  'dcp_rwcdsexplanation',
-  'dcp_communitydistrict',
-  'dcp_howdidyoudeterminethenoactionscenario',
-  'dcp_name',
-  'dcp_isrwcdsscenario',
-  'timezoneruleversionnumber',
-  'dcp_howdidyoudeterminethiswithactionscena',
-  'dcp_buildyear',
-  'dcp_developmentsiteassumptions',
-  'dcp_constructionphasing',
-  'dcp_date',
-  'overriddencreatedon',
-  'utcconversiontimezonecode',
-];
+import { RWCDS_FORM_ATTRS } from './rwcds-form.attrs';
 
 @UseInterceptors(new JsonApiSerializeInterceptor('rwcds-forms', {
   attributes: [
