+++ server/src/packages/rwcds-form/rwcds-form.service.spec.ts
@@ -2,6 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { CrmModule } from '../../crm/crm.module';
 import { RwcdsFormService } from './rwcds-form.service';
 import { CrmService } from '../../crm/crm.service';
+import { PasFormService } from '../pas-form/pas-form.service';
 
 describe('RwcdsFormService', () => {
   let rwcdsFormService: RwcdsFormService;
@@ -10,7 +11,7 @@ describe('RwcdsFormService', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [CrmModule],
-      providers: [RwcdsFormService],
+      providers: [RwcdsFormService, PasFormService],
     }).compile();
 
     rwcdsFormService = module.get<RwcdsFormService>(RwcdsFormService);
